octagon-staffing-app/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.12
azure-identity==1.17.1
azure-storage-blob==12.22.0
azure-ai-documentintelligence==1.0.0b3
azure-search-documents==11.6.0b9
openai==1.47.0
httpx==0.27.2
aiofiles==24.1.0
structlog==24.4.0
streamlit==1.39.0
PyPDF2==3.0.1
python-dotenv==1.0.1
aiohttp==3.9.5


